rendervulkan: Move CVulkanCmdBuffer to hpp
